entrega-final: agrego envio de mails y vistas de admin

+++ Entregas-Proyecto/proyecto-entrega-final/src/listeners/socketProducts.js
@@ -1,8 +1,5 @@
-//import {ProductManager} from "../dao/filesystem/productManager.js";
-//const manager = new ProductManager("./src/data/productos.json");
-import ProductServiceDao from "../dao/mongodb/products.service.js";
-const productService = new ProductServiceDao();
-
+import mongoose from "mongoose";
+import { productService } from "../repository/index.js";
 
 const socketProducts = (socketServer) => {
     socketServer.on("connection", async (socket) => {
@@ -10,27 +7,31 @@ const socketProducts = (socketServer) => {
 
         //Ver lista productos
         const listadeproductos = await productService.getProducts();
-        socket.emit("real-products",listadeproductos);
+        socket.emit("real-products", listadeproductos);
 
         //Ingresar nuevo producto
         socket.on("create-prod", async (data) => {
             console.log("Data ingresada:", data)
             const res = await productService.addProduct(data);
             console.log("Respuesta:", res);
-            
+
             const listadeproductos = await productService.getProducts();
-            socket.emit("real-products",listadeproductos);
+            socket.emit("real-products", listadeproductos);
         })
 
         //Eliminar producto
         socket.on("delete-prod", async (pid) => {
             console.log("Id ingresado:", pid);
-            
+            if (pid && !mongoose.Types.ObjectId.isValid(pid)) {
+                socket.emit("error", { error: "Id no valido" });
+                return;
+            }
+
             const res = await productService.deleteProduct(pid);
             console.log("Respuesta:", res);
-            
+
             const listadeproductos = await productService.getProducts();
-            socket.emit("real-products",listadeproductos)
+            socket.emit("real-products", listadeproductos)
         })
 
     });
